+++ lib/tickets/ui/components/read_only_banner.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:titan/l10n/app_localizations.dart';
+
+class ReadOnlyBanner extends StatelessWidget {
+  const ReadOnlyBanner({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    final l10n = AppLocalizations.of(context)!;
+    return Container(
+      width: double.infinity,
+      padding: const EdgeInsets.all(10),
+      margin: const EdgeInsets.only(bottom: 12),
+      decoration: BoxDecoration(
+        color: Colors.grey.shade200,
+        borderRadius: BorderRadius.circular(8),
+      ),
+      child: Text(
+        l10n.ticketsReadOnlyDueSales,
+        style: TextStyle(fontSize: 12, color: Colors.grey.shade700),
+      ),
+    );
+  }
+}
