+++ lib/tickets/ui/pages/edit_ticket_event_page.dart
@@ -9,12 +9,14 @@ import 'package:titan/navigation/ui/scroll_to_hide_navbar.dart';
 import 'package:titan/tickets/class/answer_type.dart';
 import 'package:titan/tickets/class/category.dart';
 import 'package:titan/tickets/class/question.dart';
-import 'package:titan/tickets/class/session.dart';
 import 'package:titan/tickets/class/ticket_event.dart';
 import 'package:titan/tickets/providers/selected_ticket_event_provider.dart';
 import 'package:titan/tickets/providers/store_tickets_list_provider.dart';
 import 'package:titan/tickets/providers/ticket_event_edit_provider.dart';
 import 'package:titan/tickets/providers/ticket_event_provider.dart';
+import 'package:titan/tickets/ui/components/edit_ticket_event_helpers.dart';
+import 'package:titan/tickets/ui/components/read_only_banner.dart';
+import 'package:titan/tickets/ui/components/sessions_section.dart';
 import 'package:titan/tickets/ui/components/ticket_event_status_chip.dart';
 import 'package:titan/tickets/ui/tickets_module.dart';
 import 'package:titan/tools/constants.dart';
@@ -24,7 +26,6 @@ import 'package:titan/tools/token_expire_wrapper.dart';
 import 'package:titan/tools/ui/builders/async_child.dart';
 import 'package:titan/tools/ui/styleguide/bottom_modal_template.dart';
 import 'package:titan/tools/ui/styleguide/button.dart';
-import 'package:titan/tools/ui/styleguide/confirm_modal.dart';
 import 'package:titan/tools/ui/widgets/date_entry.dart';
 import 'package:titan/tools/ui/widgets/text_entry.dart';
 
@@ -74,7 +75,7 @@ class _EditTicketEventContent extends HookConsumerWidget {
         return;
       }
 
-      if (!await _showDeleteConfirm(context, ref)) {
+      if (!await showDeleteConfirm(context, ref)) {
         return;
       }
       if (!context.mounted) return;
@@ -95,7 +96,7 @@ class _EditTicketEventContent extends HookConsumerWidget {
           displayToast(context, TypeMsg.msg, l10n.ticketsDeleteEventSuccess);
           QR.back();
         } else {
-          _showEditError(
+          showEditError(
             context,
             ref,
             l10n,
@@ -142,7 +143,7 @@ class _EditTicketEventContent extends HookConsumerWidget {
               style: TextStyle(fontSize: 13, color: ColorConstants.onTertiary),
             ),
             const SizedBox(height: 24),
-            _SectionCard(
+            SectionCard(
               title: l10n.ticketsGeneralInfo,
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
@@ -203,7 +204,7 @@ class _EditTicketEventContent extends HookConsumerWidget {
                     child: Text(l10n.ticketsEdit),
                   ),
                   const SizedBox(height: 16),
-                  if (!canDeleteEvent) const _ReadOnlyBanner(),
+                  if (!canDeleteEvent) const ReadOnlyBanner(),
                   SizedBox(
                     width: double.infinity,
                     child: OutlinedButton(
@@ -218,7 +219,7 @@ class _EditTicketEventContent extends HookConsumerWidget {
               ),
             ),
             const SizedBox(height: 16),
-            _SessionsSection(event: event, onEventUpdated: onEventUpdated),
+            SessionsSection(event: event, onEventUpdated: onEventUpdated),
             const SizedBox(height: 16),
             _CategoriesSection(event: event, onEventUpdated: onEventUpdated),
             const SizedBox(height: 16),
@@ -231,119 +232,9 @@ class _EditTicketEventContent extends HookConsumerWidget {
   }
 }
 
-class _SectionCard extends StatelessWidget {
-  final String title;
-  final Widget child;
-
-  const _SectionCard({required this.title, required this.child});
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      margin: EdgeInsets.zero,
-      elevation: 0,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(12),
-        side: BorderSide(
-          color: ColorConstants.secondary.withValues(alpha: 0.5),
-        ),
-      ),
-      child: Padding(
-        padding: const EdgeInsets.all(16),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              title,
-              style: const TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-            ),
-            const SizedBox(height: 16),
-            child,
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-class _ReadOnlyBanner extends StatelessWidget {
-  const _ReadOnlyBanner();
-
-  @override
-  Widget build(BuildContext context) {
-    final l10n = AppLocalizations.of(context)!;
-    return Container(
-      width: double.infinity,
-      padding: const EdgeInsets.all(10),
-      margin: const EdgeInsets.only(bottom: 12),
-      decoration: BoxDecoration(
-        color: Colors.grey.shade200,
-        borderRadius: BorderRadius.circular(8),
-      ),
-      child: Text(
-        l10n.ticketsReadOnlyDueSales,
-        style: TextStyle(fontSize: 12, color: Colors.grey.shade700),
-      ),
-    );
-  }
-}
-
-Future<bool> _showDeleteConfirm(BuildContext context, WidgetRef ref) async {
-  final l10n = AppLocalizations.of(context)!;
-  var confirmed = false;
-  await showCustomBottomModal(
-    context: context,
-    ref: ref,
-    modal: ConfirmModal(
-      title: l10n.ticketsDeleteConfirm,
-      description: l10n.globalIrreversibleAction,
-      onYes: () => confirmed = true,
-    ),
-  );
-  return confirmed;
-}
-
-void _showEditError(
-  BuildContext context,
-  WidgetRef ref,
-  AppLocalizations l10n, {
-  String? fallbackDueSales,
-  String? fallbackDueAnswers,
-}) {
-  final error = ref.read(ticketEventEditProvider).error;
-  if (error is AppException) {
-    final message = error.message;
-    if (message.contains('checkouts or tickets') && fallbackDueSales != null) {
-      displayToast(context, TypeMsg.error, fallbackDueSales);
-      return;
-    }
-    if (message.contains('answers') && fallbackDueAnswers != null) {
-      displayToast(context, TypeMsg.error, fallbackDueAnswers);
-      return;
-    }
-    displayToast(context, TypeMsg.error, message);
-    return;
-  }
-  displayToast(context, TypeMsg.error, l10n.ticketsUpdateError);
-}
-
 void _updateTicketEventState(WidgetRef ref, TicketEvent updated) {
   ref.read(ticketEventByIdProvider(updated.id).notifier).setEvent(updated);
   ref.read(selectedTicketEventProvider.notifier).state = updated;
-}
-
-TicketEvent _withSession(TicketEvent event, Session session) {
-  return event.copyWith(
-    sessions: event.sessions
-        .map((s) => s.id == session.id ? session : s)
-        .toList(),
-  );
-}
-
-TicketEvent _withoutSession(TicketEvent event, String sessionId) {
-  return event.copyWith(
-    sessions: event.sessions.where((s) => s.id != sessionId).toList(),
-  );
 }
 
 TicketEvent _withCategory(TicketEvent event, Category category) {
@@ -509,7 +400,7 @@ Future<void> _showGeneralInfoEditDialog(
                         );
                         onEventUpdated(updated);
                       } else {
-                        _showEditError(context, ref, l10n);
+                        showEditError(context, ref, l10n);
                       }
                     });
                   } catch (e) {
@@ -549,33 +440,7 @@ Future<void> _toggleEventDisabled(
       displayToast(context, TypeMsg.msg, l10n.ticketsEditSuccess);
       onEventUpdated(updated);
     } else {
-      _showEditError(context, ref, l10n);
-    }
-  });
-}
-
-Future<void> _toggleSessionDisabled(
-  BuildContext context,
-  WidgetRef ref,
-  TicketEvent event,
-  Session session,
-  bool disabled,
-  void Function(TicketEvent) onEventUpdated,
-) async {
-  final l10n = AppLocalizations.of(context)!;
-  final editNotifier = ref.read(ticketEventEditProvider.notifier);
-  await tokenExpireWrapper(ref, () async {
-    final success = await editNotifier.updateSessionDisabled(
-      event.id,
-      session.id,
-      disabled,
-    );
-    if (!context.mounted) return;
-    if (success) {
-      displayToast(context, TypeMsg.msg, l10n.ticketsEditSuccess);
-      onEventUpdated(_withSession(event, session.copyWith(disabled: disabled)));
-    } else {
-      _showEditError(context, ref, l10n);
+      showEditError(context, ref, l10n);
     }
   });
 }
@@ -603,7 +468,7 @@ Future<void> _toggleCategoryDisabled(
         _withCategory(event, category.copyWith(disabled: disabled)),
       );
     } else {
-      _showEditError(context, ref, l10n);
+      showEditError(context, ref, l10n);
     }
   });
 }
@@ -631,361 +496,9 @@ Future<void> _toggleQuestionDisabled(
         _withQuestion(event, question.copyWith(disabled: disabled)),
       );
     } else {
-      _showEditError(context, ref, l10n);
+      showEditError(context, ref, l10n);
     }
   });
-}
-
-class _SessionsSection extends HookConsumerWidget {
-  final TicketEvent event;
-  final void Function(TicketEvent) onEventUpdated;
-
-  const _SessionsSection({required this.event, required this.onEventUpdated});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final l10n = AppLocalizations.of(context)!;
-    final locale = Localizations.localeOf(context);
-    final dateFormatter = DateFormat('dd/MM/yyyy HH:mm', locale.toString());
-    final editNotifier = ref.watch(ticketEventEditProvider.notifier);
-
-    Future<void> addSession() async {
-      final nameController = TextEditingController();
-      final dateController = TextEditingController();
-      final quotaController = TextEditingController();
-
-      final confirmed = await showDialog<bool>(
-        context: context,
-        builder: (ctx) => AlertDialog(
-          title: Text(l10n.ticketsAddSession),
-          content: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              TextEntry(
-                label: l10n.ticketsSessionLabelNumbered(
-                  event.sessions.length + 1,
-                ),
-                controller: nameController,
-                onChanged: (_) {},
-              ),
-              const SizedBox(height: 8),
-              DateEntry(
-                label: l10n.ticketsDateLabel,
-                controller: dateController,
-                onTap: () => getFullDate(ctx, dateController),
-              ),
-              const SizedBox(height: 8),
-              TextEntry(
-                label: l10n.ticketsQuotaLabel,
-                controller: quotaController,
-                keyboardType: TextInputType.number,
-                canBeEmpty: true,
-                onChanged: (_) {},
-              ),
-            ],
-          ),
-          actions: [
-            TextButton(
-              onPressed: () => Navigator.pop(ctx, false),
-              child: Text(l10n.globalCancel),
-            ),
-            FilledButton(
-              onPressed: () => Navigator.pop(ctx, true),
-              child: Text(l10n.ticketsSave),
-            ),
-          ],
-        ),
-      );
-
-      if (confirmed != true || !context.mounted) return;
-      if (nameController.text.trim().isEmpty || dateController.text.isEmpty) {
-        return;
-      }
-
-      final session = Session(
-        id: '',
-        name: nameController.text.trim(),
-        startDatetime: DateTime.parse(
-          processDateBackWithHourMaybe(dateController.text, locale.toString()),
-        ),
-        quota: quotaController.text.isEmpty
-            ? null
-            : int.tryParse(quotaController.text),
-      );
-
-      await tokenExpireWrapper(ref, () async {
-        final created = await editNotifier.addSession(event.id, session);
-        if (!context.mounted) return;
-        if (created != null) {
-          displayToast(context, TypeMsg.msg, l10n.ticketsEditSuccess);
-          onEventUpdated(
-            event.copyWith(sessions: [...event.sessions, created]),
-          );
-        } else {
-          displayToast(context, TypeMsg.error, l10n.ticketsUpdateError);
-        }
-      });
-    }
-
-    return _SectionCard(
-      title: l10n.ticketsSessions,
-      child: Column(
-        children: [
-          ...event.sessions.map((session) {
-            final locked = session.hasSales;
-            return Card(
-              margin: const EdgeInsets.only(bottom: 12),
-              elevation: 0,
-              color: locked ? Colors.grey.shade100 : null,
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(8),
-                side: BorderSide(
-                  color: ColorConstants.secondary.withValues(alpha: 0.3),
-                ),
-              ),
-              child: Padding(
-                padding: const EdgeInsets.all(12),
-                child: Column(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    if (locked) const _ReadOnlyBanner(),
-                    Row(
-                      children: [
-                        Expanded(
-                          child: Text(
-                            session.name,
-                            style: const TextStyle(fontWeight: FontWeight.bold),
-                          ),
-                        ),
-                        if (session.disabled)
-                          Text(
-                            l10n.ticketsDisabled,
-                            style: TextStyle(
-                              color: Colors.grey.shade600,
-                              fontSize: 12,
-                            ),
-                          ),
-                      ],
-                    ),
-                    const SizedBox(height: 4),
-                    Text(
-                      dateFormatter.format(session.startDatetime),
-                      style: TextStyle(color: ColorConstants.tertiary),
-                    ),
-                    Text(
-                      '${l10n.ticketsQuotaLabel}: ${session.quota?.toString() ?? l10n.ticketsUnlimited}',
-                      style: TextStyle(
-                        fontSize: 12,
-                        color: ColorConstants.onTertiary,
-                      ),
-                    ),
-                    Text(
-                      '${session.ticketsSold} ${l10n.ticketsTicketsSold.toLowerCase()}, ${session.ticketsInCheckout} ${l10n.ticketsTicketsInCheckout.toLowerCase()}',
-                      style: TextStyle(
-                        fontSize: 12,
-                        color: ColorConstants.onTertiary,
-                      ),
-                    ),
-                    SwitchListTile(
-                      value: !session.disabled,
-                      onChanged: (value) => _toggleSessionDisabled(
-                        context,
-                        ref,
-                        event,
-                        session,
-                        !value,
-                        onEventUpdated,
-                      ),
-                      title: Text(
-                        session.disabled
-                            ? l10n.ticketsSessionDeactivated
-                            : l10n.ticketsSessionActivated,
-                      ),
-                      contentPadding: EdgeInsets.zero,
-                    ),
-                    if (!locked) ...[
-                      const SizedBox(height: 8),
-                      Row(
-                        children: [
-                          OutlinedButton(
-                            onPressed: () => _showSessionEditDialog(
-                              context,
-                              ref,
-                              event,
-                              session,
-                              onEventUpdated,
-                            ),
-                            child: Text(l10n.ticketsEdit),
-                          ),
-                          const SizedBox(width: 8),
-                          OutlinedButton(
-                            onPressed: () async {
-                              if (!await _showDeleteConfirm(context, ref)) {
-                                return;
-                              }
-                              if (!context.mounted) return;
-                              await tokenExpireWrapper(ref, () async {
-                                final success = await editNotifier
-                                    .deleteSession(event.id, session.id);
-                                if (!context.mounted) return;
-                                if (success) {
-                                  displayToast(
-                                    context,
-                                    TypeMsg.msg,
-                                    l10n.ticketsEditSuccess,
-                                  );
-                                  onEventUpdated(
-                                    _withoutSession(event, session.id),
-                                  );
-                                } else {
-                                  _showEditError(
-                                    context,
-                                    ref,
-                                    l10n,
-                                    fallbackDueSales:
-                                        l10n.ticketsCannotDeleteDueSales,
-                                  );
-                                }
-                              });
-                            },
-                            style: OutlinedButton.styleFrom(
-                              foregroundColor: ColorConstants.error,
-                            ),
-                            child: Text(l10n.ticketsDelete),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ],
-                ),
-              ),
-            );
-          }),
-          OutlinedButton.icon(
-            onPressed: addSession,
-            icon: const HeroIcon(HeroIcons.plus, size: 20),
-            label: Text(l10n.ticketsAddSession),
-          ),
-        ],
-      ),
-    );
-  }
-
-  Future<void> _showSessionEditDialog(
-    BuildContext context,
-    WidgetRef ref,
-    TicketEvent event,
-    Session session,
-    void Function(TicketEvent) onEventUpdated,
-  ) async {
-    final l10n = AppLocalizations.of(context)!;
-    final locale = Localizations.localeOf(context);
-    final dateFormatter = DateFormat('dd/MM/yyyy HH:mm', locale.toString());
-    final editNotifier = ref.read(ticketEventEditProvider.notifier);
-
-    final nameController = TextEditingController(text: session.name);
-    final dateController = TextEditingController(
-      text: dateFormatter.format(session.startDatetime),
-    );
-    final quotaController = TextEditingController(
-      text: session.quota?.toString() ?? '',
-    );
-    var disabled = session.disabled;
-
-    await showCustomBottomModal(
-      context: context,
-      ref: ref,
-      modal: StatefulBuilder(
-        builder: (modalContext, setState) => BottomModalTemplate(
-          title: l10n.ticketsEdit,
-          child: Column(
-            children: [
-              TextEntry(
-                label: l10n.ticketsSessionLabel,
-                controller: nameController,
-                onChanged: (_) {},
-              ),
-              const SizedBox(height: 8),
-              DateEntry(
-                label: l10n.ticketsDateLabel,
-                controller: dateController,
-                onTap: () => getFullDate(modalContext, dateController),
-              ),
-              const SizedBox(height: 8),
-              TextEntry(
-                label: l10n.ticketsQuotaLabel,
-                controller: quotaController,
-                keyboardType: TextInputType.number,
-                canBeEmpty: true,
-                onChanged: (_) {},
-              ),
-              SwitchListTile(
-                value: !disabled,
-                onChanged: (v) => setState(() => disabled = !v),
-                title: Text(
-                  disabled
-                      ? l10n.ticketsSessionDeactivated
-                      : l10n.ticketsSessionActivated,
-                ),
-                contentPadding: EdgeInsets.zero,
-              ),
-              const SizedBox(height: 20),
-              Button(
-                text: l10n.ticketsSaveChanges,
-                onPressed: () async {
-                  Navigator.pop(modalContext);
-                  if (!context.mounted) return;
-
-                  final updated = session.copyWith(
-                    name: nameController.text.trim(),
-                    startDatetime: DateTime.parse(
-                      processDateBackWithHourMaybe(
-                        dateController.text,
-                        locale.toString(),
-                      ),
-                    ),
-                    quota: quotaController.text.isEmpty
-                        ? null
-                        : int.tryParse(quotaController.text),
-                    disabled: disabled,
-                  );
-
-                  await tokenExpireWrapper(ref, () async {
-                    final success = await editNotifier.updateSession(
-                      event.id,
-                      updated,
-                    );
-                    if (!context.mounted) return;
-                    if (success) {
-                      displayToast(
-                        context,
-                        TypeMsg.msg,
-                        l10n.ticketsEditSuccess,
-                      );
-                      onEventUpdated(_withSession(event, updated));
-                    } else {
-                      final error = ref.read(ticketEventEditProvider).error;
-                      displayToast(
-                        context,
-                        TypeMsg.error,
-                        error is AppException
-                            ? error.message
-                            : l10n.ticketsUpdateError,
-                      );
-                    }
-                  });
-                },
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-
-    nameController.dispose();
-    dateController.dispose();
-    quotaController.dispose();
-  }
 }
 
 class _CategoriesSection extends HookConsumerWidget {
@@ -1079,7 +592,7 @@ class _CategoriesSection extends HookConsumerWidget {
       });
     }
 
-    return _SectionCard(
+    return SectionCard(
       title: l10n.ticketsCategories,
       child: Column(
         children: [
@@ -1100,7 +613,7 @@ class _CategoriesSection extends HookConsumerWidget {
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
-                    if (locked) const _ReadOnlyBanner(),
+                    if (locked) const ReadOnlyBanner(),
                     Row(
                       children: [
                         Expanded(
@@ -1174,7 +687,7 @@ class _CategoriesSection extends HookConsumerWidget {
                           const SizedBox(width: 8),
                           OutlinedButton(
                             onPressed: () async {
-                              if (!await _showDeleteConfirm(context, ref)) {
+                              if (!await showDeleteConfirm(context, ref)) {
                                 return;
                               }
                               if (!context.mounted) return;
@@ -1192,7 +705,7 @@ class _CategoriesSection extends HookConsumerWidget {
                                     _withoutCategory(event, category.id),
                                   );
                                 } else {
-                                  _showEditError(
+                                  showEditError(
                                     context,
                                     ref,
                                     l10n,
@@ -1462,12 +975,12 @@ class _QuestionsSection extends HookConsumerWidget {
             event.copyWith(questions: [...event.questions, created]),
           );
         } else {
-          _showEditError(context, ref, l10n);
+          showEditError(context, ref, l10n);
         }
       });
     }
 
-    return _SectionCard(
+    return SectionCard(
       title: l10n.ticketsQuestions,
       child: Column(
         children: [
@@ -1561,7 +1074,7 @@ class _QuestionsSection extends HookConsumerWidget {
                         const SizedBox(width: 8),
                         OutlinedButton(
                           onPressed: () async {
-                            if (!await _showDeleteConfirm(context, ref)) return;
+                            if (!await showDeleteConfirm(context, ref)) return;
                             if (!context.mounted) return;
                             await tokenExpireWrapper(ref, () async {
                               final success = await editNotifier.deleteQuestion(
@@ -1579,7 +1092,7 @@ class _QuestionsSection extends HookConsumerWidget {
                                   _withoutQuestion(event, question.id),
                                 );
                               } else {
-                                _showEditError(
+                                showEditError(
                                   context,
                                   ref,
                                   l10n,
@@ -1733,7 +1246,7 @@ class _QuestionsSection extends HookConsumerWidget {
                         );
                         onEventUpdated(_withQuestion(event, updatedQuestion));
                       } else {
-                        _showEditError(context, ref, l10n);
+                        showEditError(context, ref, l10n);
                       }
                     });
                   },
